# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/products": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor`). Query `tags=seafood,shrimp` uses AND semantics. Pass `lat`+`lng` (+ optional `radius`) to filter by cook proximity. Product `images` are `{ id, url }` objects.",
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive name search",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cookId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isAvailable",
            "required": false,
            "in": "query",
            "description": "Defaults to true (only available products in catalog)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "availability",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "available",
                "made_to_order"
              ]
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Comma-separated tags; AND semantics (product must include all). Normalized lowercase.",
            "schema": {
              "example": "seafood,shrimp",
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "description": "With lng+radius: proximity filter on GET /products",
            "schema": {
              "example": 3.8833,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "example": -77.0319,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Radius in meters",
            "schema": {
              "example": 10000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List products (search, category, cook, price, availability, tags AND, optional lat/lng/radius)",
        "tags": [
          "products"
        ]
      },
      "post": {
        "description": "Supports `preparationTimeHours` and optional `tags` (must exist in GET /tags catalog; create tags via POST /tags as admin). Upload images via POST /products/:id/images after create (max 5).",
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product (cook: own catalog; admin: must pass cookId)",
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/nearby": {
      "get": {
        "description": "Cursor pagination (`limit`, `cursor` encodes distance+product id). Requires `latitude` and `longitude`. Optional `tags` (comma-separated, AND), `categoryId`, `minPrice`/`maxPrice`.",
        "operationId": "ProductsController_nearby",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": 3.8833,
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "example": -77.0319,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "Radius in meters",
            "schema": {
              "default": 10000,
              "example": 10000,
              "type": "number"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Comma-separated tags; AND semantics",
            "schema": {
              "example": "seafood,shrimp",
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Products near a geo point (tags AND, category, price; sorted by cook distance)",
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "description": "Images are `{ id, url }` (upload via POST /products/:id/images).",
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get product by id",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product (owner cook or admin; isActive admin-only)",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete product (owner cook or admin)",
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{id}/images": {
      "post": {
        "description": "multipart field `file`. JPEG/PNG/WEBP, max 5 MB. Owner cook or admin only. Returns `{ id, url }`.",
        "operationId": "ProductsController_uploadImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG, PNG, or WEBP image (max 5 MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded image id and public URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageUploadResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload a product image (max 5 per product)",
        "tags": [
          "products"
        ]
      }
    },
    "/api/products/{id}/images/{imageId}": {
      "delete": {
        "description": "Removes the Cloudinary asset and the image subdocument. `imageId` comes from product.images[].id.",
        "operationId": "ProductsController_removeImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageDeletedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a product image",
        "tags": [
          "products"
        ]
      }
    },
...
    "/api/tags": {
      "get": {
        "description": "Cursor pagination ordered by `text` ASC. Optional `search` (normalized) matches tag text. Spec 009.",
        "operationId": "TagsController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive partial text match (fields depend on the resource).",
            "schema": {
              "example": "maría",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List tags (global catalog)",
        "tags": [
          "tags"
        ]
      },
      "post": {
        "description": "Normalizes text (trim + lowercase). Returns **201** when created, **200** when the tag already exists. Unique index on `text`.",
        "operationId": "TagsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create tag (admin, idempotent)",
        "tags": [
          "tags"
        ]
      }
    },
...
    "/api/admin/products/{id}": {
      "patch": {
        "operationId": "AdminController_setProductActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetActiveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate / deactivate a product",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Petronio API",
    "description": "Marketplace API for cooks and customers (Pacific region).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ImageDeletedResponseDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "example": true
          },
          "id": {
            "type": "string",
            "example": "68af1a2b3c4d5e6f78901234",
            "description": "Present for product image deletes"
          }
        },
        "required": [
          "deleted"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Encocado de camarón"
          },
          "description": {
            "type": "string",
            "example": "Traditional Pacific shrimp stew"
          },
          "price": {
            "type": "number",
            "example": 35000
          },
          "stock": {
            "type": "number",
            "example": 8,
            "description": "Stock units (used when availability=available)"
          },
          "categoryId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "cookId": {
            "type": "string",
            "description": "Required for admin; ignored for cooks (uses their profile)",
            "example": "507f1f77bcf86cd799439011"
          },
          "availability": {
            "type": "string",
            "enum": [
              "available",
              "made_to_order"
            ],
            "example": "available"
          },
          "preparationTimeHours": {
            "type": "number",
            "example": 48,
            "description": "Preparation time in hours (e.g. 48 = 2 days)"
          },
          "minimumOrderQuantity": {
            "type": "number",
            "example": 5,
            "description": "Required minimum when made_to_order"
          },
          "tags": {
            "example": [
              "seafood",
              "shrimp",
              "traditional",
              "pacific-food"
            ],
            "description": "Must already exist in the global tag catalog (POST /tags). Normalized to lowercase; max 10; unique",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isAvailable": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "price",
          "availability"
        ]
      },
      "ProductImageUploadResponseDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://res.cloudinary.com/demo/image/upload/v1/sample.jpg"
          },
          "id": {
            "type": "string",
            "example": "68af1a2b3c4d5e6f78901234",
            "description": "Subdocument id — use with DELETE /products/:id/images/:imageId"
          }
        },
        "required": [
          "url",
          "id"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          },
          "categoryId": {
            "type": "string"
          },
          "availability": {
            "type": "string",
            "enum": [
              "available",
              "made_to_order"
            ]
          },
          "preparationTimeHours": {
            "type": "number",
            "example": 24,
            "description": "Preparation time in hours"
          },
          "minimumOrderQuantity": {
            "type": "number"
          },
          "tags": {
            "example": [
              "seafood",
              "shrimp"
            ],
            "description": "Must already exist in the global tag catalog (POST /tags). Normalized to lowercase; max 10; unique",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isAvailable": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean",
            "description": "Admin soft-deactivate"
          }
        }
      },
...
      "TagResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "68af1a2b3c4d5e6f78901234"
          },
          "text": {
            "type": "string",
            "example": "seafood"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-16T20:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-16T20:00:00.000Z"
          }
        },
        "required": [
          "id",
          "text"
        ]
      },
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "object",
            "nullable": true,
            "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ"
          },
          "hasMore": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "hasMore"
        ]
      },
      "TagsListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
      "CreateTagDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "example": "Seafood",
            "maxLength": 50
          }
        },
        "required": [
          "text"
        ]
      },
...
      "SetActiveDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "isActive"
        ]
      }
    }
  }
}